service/pipeline/requirements.txt
numpy==1.23.0
pillow==9.2.0
requests==2.28.1
uvicorn==0.13.3
nest_asyncio==1.4.3
starlette==0.22.0
python_gitlab==3.5.0
fastapi==0.89.1
stress-injector
httplib2==0.22.0
pymongo==4.5.0
psutil>=5.9.0
tqdm>=4.56.0
scipy==1.10.1
tensorflow
transformers
sentencepiece
diffusers
transformers==4.34.1
accelerate
joblib
py7zr==0.20.6
google-cloud-storage==2.12.0